:fire: fix: move coins to sentio from const folder
