+++ views/metrics/metrics.utils.ts
@@ -1,6 +1,7 @@
-import { CoinInfo, COINS_MAP } from 'interface/sentio';
+import { CoinInfo } from 'interface/sentio';
 
 import { Network } from '@/constants';
+import { COINS_MAP } from '@/constants/coins';
 import { RECOMMENDED_POOLS } from '@/constants/pools';
 
 import { PoolCardProps as IPool } from '../pools/pool-card/pool-card.types';
